refactor libevm sources
