+++ libevm/lib/database.go
