+++ libevm/lib/database_test.go
