+++ libevm/lib/hash.go
