+++ libevm/lib/hash_test.go
